near-vm-runner: move VMConfig to the most appropriate crate
VMConfig, VMLimitConfig and similar types are a public interface of the
`near-vm-runner` crate rather than the `near-primitives-core` or any
other, and as such should be exported from `near-vm-runner`.

As a more pragmatic reasoning: with limited replayability we want
near-vm-runner to be as self-contained as possible, and having types
that are somewhat of an implementation detail of itself in an external
crate is just asking for trouble.
