+++ integration-tests/src/tests/client/features/increase_deployment_cost.rs
@@ -3,12 +3,12 @@ use near_chain::ChainGenesis;
 use near_chain_configs::Genesis;
 use near_client::test_utils::TestEnv;
 use near_crypto::{InMemorySigner, KeyType};
-use near_primitives::config::VMConfig;
 use near_primitives::runtime::config_store::RuntimeConfigStore;
 use near_primitives::transaction::{Action, DeployContractAction};
 use near_primitives::version::ProtocolFeature;
 use near_primitives::views::FinalExecutionStatus;
 use near_vm_runner::internal::VMKind;
+use near_vm_runner::logic::Config as VMConfig;
 use nearcore::config::GenesisExt;
 
 use crate::tests::client::utils::TestEnvNightshadeSetupExt;
