+++ integration-tests/src/tests/runtime/sanity_checks.rs
@@ -251,11 +251,12 @@ fn test_sanity_used_gas() {
         //
         // In this test we account for this by setting `op_cost` to zero, but if future tests
         // change test WASM in significant ways, this approach may become incorrect.
-        near_primitives::config::ContractPrepareVersion::V0
-        | near_primitives::config::ContractPrepareVersion::V1 => 0,
+        near_vm_runner::ContractPrepareVersion::V0 | near_vm_runner::ContractPrepareVersion::V1 => {
+            0
+        }
         // Gas accounting is precise and instructions executed between calls to the side-effectful
         // `used_gas` host function calls will be observbable.
-        near_primitives::config::ContractPrepareVersion::V2 => {
+        near_vm_runner::ContractPrepareVersion::V2 => {
             u64::from(runtime_config.wasm_config.regular_op_cost)
         }
     };
