+++ runtime/near-vm-runner/fuzz/fuzz_targets/diffrunner.rs
@@ -6,7 +6,7 @@ use near_primitives::version::PROTOCOL_VERSION;
 use near_vm_runner::internal::VMKind;
 use near_vm_runner::logic::errors::FunctionCallError;
 use near_vm_runner::logic::mocks::mock_external::MockedExternal;
-use near_vm_runner::logic::{VMConfig, VMOutcome};
+use near_vm_runner::logic::{Config, VMOutcome};
 use near_vm_runner_fuzz::{create_context, find_entry_point, ArbitraryModule};
 
 libfuzzer_sys::fuzz_target!(|module: ArbitraryModule| {
@@ -20,7 +20,7 @@ fn run_fuzz(code: &ContractCode, vm_kind: VMKind) -> VMOutcome {
     let mut fake_external = MockedExternal::new();
     let mut context = create_context(vec![]);
     context.prepaid_gas = 10u64.pow(14);
-    let mut config = VMConfig::test();
+    let mut config = Config::test();
     config.limit_config.contract_prepare_version =
         near_vm_runner::logic::ContractPrepareVersion::V2;
     let fees = RuntimeFeesConfig::test();
