+++ runtime/near-vm-runner/src/logic/tests/ed25519_verify.rs
@@ -1,8 +1,9 @@
 use crate::logic::tests::helpers::*;
 use crate::logic::tests::vm_logic_builder::VMLogicBuilder;
 use crate::logic::HostError;
-use crate::logic::{ExtCosts, VMLogicError};
+use crate::logic::VMLogicError;
 use crate::map;
+use near_primitives_core::config::ExtCosts;
 use std::collections::HashMap;
 
 const SIGNATURE: [u8; 64] = [
