+++ runtime/near-vm-runner/src/logic/tests/gas_counter.rs
@@ -1,8 +1,8 @@
 use crate::logic::tests::helpers::*;
 use crate::logic::tests::vm_logic_builder::{TestVMLogic, VMLogicBuilder};
 use crate::logic::types::Gas;
+use crate::logic::{Config, MemSlice};
 use crate::logic::{HostError, VMLogicError};
-use crate::logic::{MemSlice, VMConfig};
 use borsh::BorshSerialize;
 use expect_test::expect;
 use near_primitives_core::config::{ActionCosts, ExtCosts};
@@ -718,7 +718,7 @@ fn create_promise_dependency(logic: &mut TestVMLogic) -> Result<(), VMLogicError
 /// Given the limit in gas, compute the corresponding limit in wasm ops for use
 /// with [`VMLogic::gas`] function.
 fn op_limit(gas_limit: Gas) -> u32 {
-    (gas_limit / (VMConfig::test().regular_op_cost as u64)) as u32
+    (gas_limit / (Config::test().regular_op_cost as u64)) as u32
 }
 
 fn test_pk() -> Vec<u8> {
