+++ runtime/near-vm-runner/src/logic/tests/logs.rs
@@ -1,8 +1,9 @@
 use crate::logic::tests::helpers::*;
 use crate::logic::tests::vm_logic_builder::VMLogicBuilder;
 use crate::logic::HostError;
-use crate::logic::{ExtCosts, MemSlice, VMLogic, VMLogicError};
+use crate::logic::{MemSlice, VMLogic, VMLogicError};
 use crate::map;
+use near_primitives_core::config::ExtCosts;
 
 #[test]
 fn test_valid_utf8() {
