+++ runtime/near-vm-runner/src/prepare.rs
@@ -3,7 +3,7 @@
 
 use crate::internal::VMKind;
 use crate::logic::errors::PrepareError;
-use crate::logic::VMConfig;
+use crate::logic::Config;
 
 mod prepare_v0;
 mod prepare_v1;
@@ -17,12 +17,12 @@ mod prepare_v2;
 /// - module doesn't define an internal memory instance,
 /// - imported memory (if any) doesn't reserve more memory than permitted by the `config`,
 /// - all imported functions from the external environment matches defined by `env` module,
-/// - functions number does not exceed limit specified in VMConfig,
+/// - functions number does not exceed limit specified in Config,
 ///
 /// The preprocessing includes injecting code for gas metering and metering the height of stack.
 pub fn prepare_contract(
     original_code: &[u8],
-    config: &VMConfig,
+    config: &Config,
     kind: VMKind,
 ) -> Result<Vec<u8>, PrepareError> {
     let prepare = config.limit_config.contract_prepare_version;
@@ -55,7 +55,7 @@ mod tests {
     use assert_matches::assert_matches;
 
     fn parse_and_prepare_wat(
-        config: &VMConfig,
+        config: &Config,
         vm_kind: VMKind,
         wat: &str,
     ) -> Result<Vec<u8>, PrepareError> {
@@ -65,7 +65,7 @@ mod tests {
 
     #[test]
     fn internal_memory_declaration() {
-        let config = VMConfig::test();
+        let config = Config::test();
         with_vm_variants(&config, |kind| {
             let r = parse_and_prepare_wat(&config, kind, r#"(module (memory 1 1))"#);
             assert_matches!(r, Ok(_));
@@ -74,7 +74,7 @@ mod tests {
 
     #[test]
     fn memory_imports() {
-        let config = VMConfig::test();
+        let config = Config::test();
 
         // This test assumes that maximum page number is configured to a certain number.
         assert_eq!(config.limit_config.max_memory_pages, 2048);
@@ -119,7 +119,7 @@ mod tests {
 
     #[test]
     fn multiple_valid_memory_are_disabled() {
-        let config = VMConfig::test();
+        let config = Config::test();
         with_vm_variants(&config, |kind| {
             // Our preparation and sanitization pass assumes a single memory, so we should fail when
             // there are multiple specified.
@@ -146,7 +146,7 @@ mod tests {
 
     #[test]
     fn imports() {
-        let config = VMConfig::test();
+        let config = Config::test();
         with_vm_variants(&config, |kind| {
             // nothing can be imported from non-"env" module for now.
             let r = parse_and_prepare_wat(
