+++ runtime/near-vm-runner/src/tests/ts_contract.rs
@@ -1,7 +1,7 @@
 use crate::logic::errors::{FunctionCallError, HostError};
 use crate::logic::mocks::mock_external::MockedExternal;
 use crate::logic::types::ReturnData;
-use crate::logic::{External, StorageGetMode, VMConfig};
+use crate::logic::{Config, External, StorageGetMode};
 use near_primitives_core::contract::ContractCode;
 use near_primitives_core::runtime::fees::RuntimeFeesConfig;
 
@@ -10,7 +10,7 @@ use crate::vm_kind::VMKind;
 
 #[test]
 pub fn test_ts_contract() {
-    let config = VMConfig::test();
+    let config = Config::test();
     with_vm_variants(&config, |vm_kind: VMKind| {
         let code = ContractCode::new(near_test_contracts::ts_contract().to_vec(), None);
         let mut fake_external = MockedExternal::new();
