+++ runtime/runtime-params-estimator/src/costs_to_runtime_config.rs
@@ -1,8 +1,9 @@
+use near_primitives::config::{ActionCosts, ExtCosts, ExtCostsConfig, ParameterCost};
 use near_primitives::runtime::config::AccountCreationConfig;
 use near_primitives::runtime::config_store::RuntimeConfigStore;
 use near_primitives::runtime::fees::{Fee, RuntimeFeesConfig};
 use near_primitives::version::PROTOCOL_VERSION;
-use near_vm_runner::logic::{ActionCosts, ExtCosts, ExtCostsConfig, ParameterCost, VMConfig};
+use near_vm_runner::logic::{Config as VMConfig, StorageGetMode};
 use node_runtime::config::RuntimeConfig;
 
 use anyhow::Context;
@@ -36,7 +37,7 @@ pub fn costs_to_runtime_config(cost_table: &CostTable) -> anyhow::Result<Runtime
             regular_op_cost: u32::try_from(regular_op_cost).unwrap(),
             disable_9393_fix: false,
             limit_config: vm_limit_config,
-            flat_storage_reads: true,
+            storage_get_mode: StorageGetMode::FlatStorage,
             fix_contract_loading_cost: false,
             implicit_account_creation: true,
         },
