+++ runtime/runtime-params-estimator/src/utils.rs
@@ -2,11 +2,12 @@ use crate::apply_block_cost;
 use crate::estimator_context::EstimatorContext;
 use crate::gas_cost::{GasCost, NonNegativeTolerance};
 use crate::transaction_builder::TransactionBuilder;
+use near_primitives::config::ExtCosts;
 use near_primitives::transaction::{
     Action, DeployContractAction, FunctionCallAction, SignedTransaction,
 };
 use near_vm_runner::internal::VMKind;
-use near_vm_runner::logic::{ExtCosts, VMConfig};
+use near_vm_runner::logic::Config as VMConfig;
 use rand::distributions::Alphanumeric;
 use rand::Rng;
 use rand_xorshift::XorShiftRng;
