+++ runtime/runtime/src/verifier.rs
@@ -1,32 +1,28 @@
-use near_crypto::key_conversion::is_valid_staking_key;
-use near_primitives::action::delegate::SignedDelegateAction;
-use near_primitives::checked_feature;
-use near_primitives::runtime::config::RuntimeConfig;
-use near_primitives::transaction::DeleteAccountAction;
-use near_primitives::types::{BlockHeight, StorageUsage};
-use near_primitives::version::ProtocolFeature;
-use near_primitives::{
-    account::AccessKeyPermission,
-    config::VMLimitConfig,
-    errors::{
-        ActionsValidationError, InvalidAccessKeyError, InvalidTxError, ReceiptValidationError,
-        RuntimeError,
-    },
-    receipt::{ActionReceipt, DataReceipt, Receipt, ReceiptEnum},
-    transaction::{
-        Action, AddKeyAction, DeployContractAction, FunctionCallAction, SignedTransaction,
-        StakeAction,
-    },
-    types::{AccountId, Balance},
-    version::ProtocolVersion,
-};
-use near_store::{
-    get_access_key, get_account, set_access_key, set_account, StorageError, TrieUpdate,
-};
-
 use crate::config::{total_prepaid_gas, tx_cost, TransactionCost};
 use crate::near_primitives::account::Account;
 use crate::VerificationResult;
+use near_crypto::key_conversion::is_valid_staking_key;
+use near_primitives::account::AccessKeyPermission;
+use near_primitives::action::delegate::SignedDelegateAction;
+use near_primitives::checked_feature;
+use near_primitives::errors::{
+    ActionsValidationError, InvalidAccessKeyError, InvalidTxError, ReceiptValidationError,
+    RuntimeError,
+};
+use near_primitives::receipt::{ActionReceipt, DataReceipt, Receipt, ReceiptEnum};
+use near_primitives::runtime::config::RuntimeConfig;
+use near_primitives::transaction::DeleteAccountAction;
+use near_primitives::transaction::{
+    Action, AddKeyAction, DeployContractAction, FunctionCallAction, SignedTransaction, StakeAction,
+};
+use near_primitives::types::{AccountId, Balance};
+use near_primitives::types::{BlockHeight, StorageUsage};
+use near_primitives::version::ProtocolFeature;
+use near_primitives::version::ProtocolVersion;
+use near_store::{
+    get_access_key, get_account, set_access_key, set_account, StorageError, TrieUpdate,
+};
+use near_vm_runner::logic::LimitConfig;
 
 pub const ZERO_BALANCE_ACCOUNT_STORAGE_LIMIT: StorageUsage = 770;
 
@@ -280,7 +276,7 @@ pub fn verify_and_charge_transaction(
 
 /// Validates a given receipt. Checks validity of the Action or Data receipt.
 pub(crate) fn validate_receipt(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     receipt: &Receipt,
     current_protocol_version: ProtocolVersion,
 ) -> Result<(), ReceiptValidationError> {
@@ -306,7 +302,7 @@ pub(crate) fn validate_receipt(
 
 /// Validates given ActionReceipt. Checks validity of the number of input data dependencies and all actions.
 fn validate_action_receipt(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     receipt: &ActionReceipt,
     current_protocol_version: ProtocolVersion,
 ) -> Result<(), ReceiptValidationError> {
@@ -322,7 +318,7 @@ fn validate_action_receipt(
 
 /// Validates given data receipt. Checks validity of the length of the returned data.
 fn validate_data_receipt(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     receipt: &DataReceipt,
 ) -> Result<(), ReceiptValidationError> {
     let data_len = receipt.data.as_ref().map(|data| data.len()).unwrap_or(0);
@@ -343,7 +339,7 @@ fn validate_data_receipt(
 /// - Validates each individual action.
 /// - Checks that the total prepaid gas doesn't exceed the limit.
 pub(crate) fn validate_actions(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     actions: &[Action],
     current_protocol_version: ProtocolVersion,
 ) -> Result<(), ActionsValidationError> {
@@ -392,7 +388,7 @@ pub(crate) fn validate_actions(
 
 /// Validates a single given action. Checks limits if applicable.
 pub fn validate_action(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     action: &Action,
     current_protocol_version: ProtocolVersion,
 ) -> Result<(), ActionsValidationError> {
@@ -410,7 +406,7 @@ pub fn validate_action(
 }
 
 fn validate_delegate_action(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     signed_delegate_action: &SignedDelegateAction,
     current_protocol_version: ProtocolVersion,
 ) -> Result<(), ActionsValidationError> {
@@ -421,7 +417,7 @@ fn validate_delegate_action(
 
 /// Validates `DeployContractAction`. Checks that the given contract size doesn't exceed the limit.
 fn validate_deploy_contract_action(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     action: &DeployContractAction,
 ) -> Result<(), ActionsValidationError> {
     if action.code.len() as u64 > limit_config.max_contract_size {
@@ -437,7 +433,7 @@ fn validate_deploy_contract_action(
 /// Validates `FunctionCallAction`. Checks that the method name length doesn't exceed the limit and
 /// the length of the arguments doesn't exceed the limit.
 fn validate_function_call_action(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     action: &FunctionCallAction,
 ) -> Result<(), ActionsValidationError> {
     if action.gas == 0 {
@@ -476,15 +472,15 @@ fn validate_stake_action(action: &StakeAction) -> Result<(), ActionsValidationEr
 /// total number of bytes of the method names doesn't exceed the limit and
 /// every method name length doesn't exceed the limit.
 fn validate_add_key_action(
-    limit_config: &VMLimitConfig,
+    limit_config: &LimitConfig,
     action: &AddKeyAction,
 ) -> Result<(), ActionsValidationError> {
     if let AccessKeyPermission::FunctionCall(fc) = &action.access_key.permission {
         // Check whether `receiver_id` is a valid account_id. Historically, we
         // allowed arbitrary strings there!
         match limit_config.account_id_validity_rules_version {
-            near_vm_runner::logic::AccountIdValidityRulesVersion::V0 => (),
-            near_vm_runner::logic::AccountIdValidityRulesVersion::V1 => {
+            near_primitives_core::config::AccountIdValidityRulesVersion::V0 => (),
+            near_primitives_core::config::AccountIdValidityRulesVersion::V1 => {
                 if let Err(_) = fc.receiver_id.parse::<AccountId>() {
                     return Err(ActionsValidationError::InvalidAccountId {
                         account_id: truncate_string(&fc.receiver_id, AccountId::MAX_LEN * 2),
@@ -1492,7 +1488,7 @@ mod tests {
 
     #[test]
     fn test_validate_receipt_valid() {
-        let limit_config = VMLimitConfig::test();
+        let limit_config = LimitConfig::test();
         validate_receipt(
             &limit_config,
             &Receipt::new_balance_refund(&alice_account(), 10),
@@ -1503,7 +1499,7 @@ mod tests {
 
     #[test]
     fn test_validate_action_receipt_too_many_input_deps() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         limit_config.max_number_input_data_dependencies = 1;
         assert_eq!(
             validate_action_receipt(
@@ -1530,7 +1526,7 @@ mod tests {
 
     #[test]
     fn test_validate_data_receipt_valid() {
-        let limit_config = VMLimitConfig::test();
+        let limit_config = LimitConfig::test();
         validate_data_receipt(
             &limit_config,
             &DataReceipt { data_id: CryptoHash::default(), data: None },
@@ -1546,7 +1542,7 @@ mod tests {
 
     #[test]
     fn test_validate_data_receipt_too_much_data() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         let data = b"hello".to_vec();
         limit_config.max_length_returned_data = data.len() as u64 - 1;
         assert_eq!(
@@ -1566,13 +1562,13 @@ mod tests {
 
     #[test]
     fn test_validate_actions_empty() {
-        let limit_config = VMLimitConfig::test();
+        let limit_config = LimitConfig::test();
         validate_actions(&limit_config, &[], PROTOCOL_VERSION).expect("empty actions");
     }
 
     #[test]
     fn test_validate_actions_valid_function_call() {
-        let limit_config = VMLimitConfig::test();
+        let limit_config = LimitConfig::test();
         validate_actions(
             &limit_config,
             &[Action::FunctionCall(Box::new(FunctionCallAction {
@@ -1588,7 +1584,7 @@ mod tests {
 
     #[test]
     fn test_validate_actions_too_much_gas() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         limit_config.max_total_prepaid_gas = 220;
         assert_eq!(
             validate_actions(
@@ -1616,7 +1612,7 @@ mod tests {
 
     #[test]
     fn test_validate_actions_gas_overflow() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         limit_config.max_total_prepaid_gas = 220;
         assert_eq!(
             validate_actions(
@@ -1644,7 +1640,7 @@ mod tests {
 
     #[test]
     fn test_validate_actions_num_actions() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         limit_config.max_actions_per_receipt = 1;
         assert_eq!(
             validate_actions(
@@ -1665,7 +1661,7 @@ mod tests {
 
     #[test]
     fn test_validate_delete_must_be_final() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         limit_config.max_actions_per_receipt = 3;
         assert_eq!(
             validate_actions(
@@ -1685,7 +1681,7 @@ mod tests {
 
     #[test]
     fn test_validate_delete_must_work_if_its_final() {
-        let mut limit_config = VMLimitConfig::test();
+        let mut limit_config = LimitConfig::test();
         limit_config.max_actions_per_receipt = 3;
         assert_eq!(
             validate_actions(
@@ -1707,7 +1703,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_create_account() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::CreateAccount(CreateAccountAction {}),
             PROTOCOL_VERSION,
         )
@@ -1717,7 +1713,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_function_call() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::FunctionCall(Box::new(FunctionCallAction {
                 method_name: "hello".to_string(),
                 args: b"abc".to_vec(),
@@ -1733,7 +1729,7 @@ mod tests {
     fn test_validate_action_invalid_function_call_zero_gas() {
         assert_eq!(
             validate_action(
-                &VMLimitConfig::test(),
+                &LimitConfig::test(),
                 &Action::FunctionCall(Box::new(FunctionCallAction {
                     method_name: "new".to_string(),
                     args: vec![],
@@ -1750,7 +1746,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_transfer() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::Transfer(TransferAction { deposit: 10 }),
             PROTOCOL_VERSION,
         )
@@ -1760,7 +1756,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_stake() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::Stake(Box::new(StakeAction {
                 stake: 100,
                 public_key: "ed25519:KuTCtARNzxZQ3YvXDeLjx83FDqxv2SdQTSbiq876zR7".parse().unwrap(),
@@ -1774,7 +1770,7 @@ mod tests {
     fn test_validate_action_invalid_staking_key() {
         assert_eq!(
             validate_action(
-                &VMLimitConfig::test(),
+                &LimitConfig::test(),
                 &Action::Stake(Box::new(StakeAction {
                     stake: 100,
                     public_key: PublicKey::empty(KeyType::ED25519),
@@ -1791,7 +1787,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_add_key_full_permission() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::AddKey(Box::new(AddKeyAction {
                 public_key: PublicKey::empty(KeyType::ED25519),
                 access_key: AccessKey::full_access(),
@@ -1804,7 +1800,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_add_key_function_call() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::AddKey(Box::new(AddKeyAction {
                 public_key: PublicKey::empty(KeyType::ED25519),
                 access_key: AccessKey {
@@ -1824,7 +1820,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_delete_key() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::DeleteKey(Box::new(DeleteKeyAction {
                 public_key: PublicKey::empty(KeyType::ED25519),
             })),
@@ -1836,7 +1832,7 @@ mod tests {
     #[test]
     fn test_validate_action_valid_delete_account() {
         validate_action(
-            &VMLimitConfig::test(),
+            &LimitConfig::test(),
             &Action::DeleteAccount(DeleteAccountAction { beneficiary_id: alice_account() }),
             PROTOCOL_VERSION,
         )
@@ -1861,7 +1857,7 @@ mod tests {
         };
         assert_eq!(
             validate_actions(
-                &VMLimitConfig::test(),
+                &LimitConfig::test(),
                 &[
                     Action::Delegate(Box::new(signed_delegate_action.clone())),
                     Action::Delegate(Box::new(signed_delegate_action.clone())),
@@ -1872,7 +1868,7 @@ mod tests {
         );
         assert_eq!(
             validate_actions(
-                &&VMLimitConfig::test(),
+                &&LimitConfig::test(),
                 &[Action::Delegate(Box::new(signed_delegate_action.clone())),],
                 PROTOCOL_VERSION,
             ),
@@ -1880,7 +1876,7 @@ mod tests {
         );
         assert_eq!(
             validate_actions(
-                &VMLimitConfig::test(),
+                &LimitConfig::test(),
                 &[
                     Action::CreateAccount(CreateAccountAction {}),
                     Action::Delegate(Box::new(signed_delegate_action)),
